requirements.txt
Flask
psutil
pandas
speedtest-cli
requests